requirements.txt
# HTML to PDF conversion - 重要なので最初に配置
xhtml2pdf==0.2.11

pdfplumber==0.9.0
python-dotenv==1.0.0
requests==2.31.0
Flask==2.2.3
Werkzeug==2.2.3
Flask-WTF==1.1.1
Flask-Session==0.5.0
Flask-Login==0.6.2
WTForms==3.0.1
paypalrestsdk==1.13.1
pytesseract==0.3.10
gunicorn==20.1.0
PyPDF2==3.0.1
Pillow
pdf2image==1.16.3
# NumPyを1.xシリーズに固定（OpenCVとの互換性のため）
numpy==1.26.4
opencv-python==4.8.0.74
pdfminer.six==20221105
google-cloud-vision==3.4.0
# az-cognitive-services removed for compatibility
boto3==1.28.38

# APIキーキャッシュ機能用のライブラリ
google-cloud-firestore==2.11.1
cryptography==41.0.7

# Local package installation removed for Cloud Build compatibility
